Don't remove the blockchain state in case lastExecutedSeqNum is zero

+++ kvbc/src/replica_state_sync_imp.cpp
@@ -1,6 +1,6 @@
 // Concord
 //
-// Copyright (c) 2019 VMware, Inc. All Rights Reserved.
+// Copyright (c) 2019-2021 VMware, Inc. All Rights Reserved.
 //
 // This product is licensed to you under the Apache 2.0 license (the
 // "License").  You may not use this product except in compliance with the
@@ -19,18 +19,19 @@
 
 using concord::kvbc::Key;
 
-namespace concord {
-namespace kvbc {
+namespace concord::kvbc {
 
 ReplicaStateSyncImp::ReplicaStateSyncImp(IBlockMetadata* blockMetadata) : blockMetadata_(blockMetadata) {}
 
-uint64_t ReplicaStateSyncImp::execute(logging::Logger& logger,
-                                      IDbAdapter& bcDBAdapter,
-                                      BlockId lastReachableBlockId,
-                                      uint64_t lastExecutedSeqNum) {
+uint64_t ReplicaStateSyncImp::execute(logging::Logger& logger, IDbAdapter& bcDBAdapter, uint64_t lastExecutedSeqNum) {
+  if (!lastExecutedSeqNum) {
+    LOG_INFO(logger, "Replica's metadata is empty => skip blocks removal");
+    return 0;
+  }
   uint64_t removedBlocksNum = 0;
   const auto genesisBlockId = bcDBAdapter.getGenesisBlockId();
   const auto blockMetadataKey = blockMetadata_->getKey();
+  BlockId lastReachableBlockId = bcDBAdapter.getLastReachableBlockId();
   uint64_t lastBlockSeqNum = 0;
   while (lastReachableBlockId && genesisBlockId <= lastReachableBlockId) {
     // Get execution sequence number stored in the current last block.
@@ -43,13 +44,19 @@ uint64_t ReplicaStateSyncImp::execute(logging::Logger& logger,
     }
     // SBFT State Metadata is not in sync with the Blockchain State.
     // Remove blocks which sequence number is greater than lastExecutedSeqNum.
+    if (removedBlocksNum == 1) {
+      std::string error = " Detected more than one block needs to be deleted from the blockchain - unsupported";
+      LOG_FATAL(logger, error);
+      throw std::runtime_error(__PRETTY_FUNCTION__ + error);
+    }
     bcDBAdapter.deleteLastReachableBlock();
-    --lastReachableBlockId;
+    lastReachableBlockId = bcDBAdapter.getLastReachableBlockId();
     ++removedBlocksNum;
   }
-  LOG_INFO(logger, "All blockchain blocks deleted " << KVLOG(removedBlocksNum, lastBlockSeqNum, lastReachableBlockId));
+  LOG_INFO(logger,
+           "Inconsistent blockchain block deleted "
+               << KVLOG(removedBlocksNum, lastExecutedSeqNum, lastBlockSeqNum, bcDBAdapter.getLastReachableBlockId()));
   return removedBlocksNum;
 }
 
-}  // namespace kvbc
-}  // namespace concord
+}  // namespace concord::kvbc
